Rename model.py to cuds.py
